move run function out of cli and into jrnl
