+++ pyproject.toml
@@ -40,7 +40,7 @@ pyflakes = "^2.2.0"
 pytest = "^5.4.3"
 
 [tool.poetry.scripts]
-jrnl = 'jrnl.cli:run'
+jrnl = 'jrnl.cli:cli'
 
 [tool.isort]
 multi_line_output = 7
